Updated button styles
Improved buttons styles, so it look a lot better.
Like button is now blue and the dislike buttom is red. Also there is some space betweens the button too, so they seperate from each other.
In addition the selected button is now gold, so when any button is selected (like/dislike), the color changes to gold, so the viewer know they have selected that option.